third_party/blink/web_tests/editing/shadow/selection-all-with-shadow-expected.txt

SelectAll and Shadow DOM Tree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Select all from host content in shadow tree => select all on document
PASS selection.anchorNode.nodeValue is "SelectAll and Shadow DOM Tree"
PASS selection.anchorOffset is 0
PASS selection.focusNode.nodeValue is "end of document"
PASS selection.focusOffset is 15

Select all from "EFO" in shadow tree => select descendants of immediate child of shadow root containing start position
PASS selection.anchorNode is host1.parentNode
PASS selection.anchorOffset is 1
PASS selection.focusNode is host1.parentNode
PASS selection.focusOffset is 1
PASS selectionInHost1.anchorNode.nodeValue is "BEFORE"
PASS selectionInHost1.anchorOffset is 0
PASS selectionInHost1.focusNode.nodeValue is "BEFORE"
PASS selectionInHost1.focusOffset is 6

Select all from host content in editable => select all in editable rather than whole document
PASS selection.anchorNode.nodeValue is "012"
PASS selection.anchorOffset is 0
PASS selection.focusNode.nodeValue is "345"
PASS selection.anchorOffset is 0

Select all from "EFO" in shadow tree => select descendants of immediate child of shadow root containing start position, host editable doesn't affect select all
PASS selection.anchorNode is host2.parentNode
PASS selection.anchorOffset is 1
PASS selection.focusNode is host2.parentNode
PASS selection.focusOffset is 1
PASS selectionInHost2.anchorNode.nodeValue is "BEFORE"
PASS selectionInHost2.anchorOffset is 0
PASS selectionInHost2.focusNode.nodeValue is "BEFORE"
PASS selectionInHost2.focusOffset is 6

Select all from host content in nested shadow tree => select all on document
PASS selection.anchorNode.nodeValue is "SelectAll and Shadow DOM Tree"
PASS selection.anchorOffset is 0
PASS selection.focusNode.nodeValue is "end of document"
PASS selection.focusOffset is 15

Select all from "EFO" in nested shadow tree => select descendants of immediate child of shadow root containing start position
PASS selection.anchorNode is host4.parentNode
PASS selection.anchorOffset is 1
PASS selection.focusNode is host4.parentNode
PASS selection.focusOffset is 1
PASS selectionInHost3.anchorNode is null
PASS selectionInHost3.anchorOffset is 0
PASS selectionInHost3.focusNode is null
PASS selectionInHost3.focusOffset is 0
PASS selectionInHost4.anchorNode.nodeValue is "BEFORE"
PASS selectionInHost4.anchorOffset is 0
PASS selectionInHost4.focusNode.nodeValue is "BEFORE"
PASS selectionInHost4.focusOffset is 6
PASS successfullyParsed is true

TEST COMPLETE

012
host1
345
012
host2
345
012
host3abc
host4
host3def
345
012
host5
345

end of document
